# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Area Backend API
  version: 1.0.0
  description: |-
    HTTP API exposed by the Area backend. All routes are mounted under `CONFIG.BASE_PATH`
    (default `/api`). Authentication uses bearer JWTs issued by `/auth/login`. Runner
    endpoints use a shared nonce/token pair carried in custom headers.
servers:
  - url: /api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by `/auth/login`
    RunnerJobId:
      type: apiKey
      in: header
      name: X-Runner-Job
      description: Runner job identifier returned when the workflow execution was created
    RunnerToken:
      type: apiKey
      in: header
      name: X-Runner-Token
      description: HMAC token used to authenticate runner -> backend callbacks
    PasswordResetToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token sent with password reset emails ("Bearer <token>")
  schemas:
...
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    UserImageUploadRequest:
      type: object
      required:
        - image
      properties:
        image:
          type: string
          format: byte
          description: Base64 encoded PNG data; data URI prefixes are accepted
...
    TeamSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    TeamListResponse:
      type: object
      properties:
        teams:
          type: array
          items:
            $ref: '#/components/schemas/TeamSummary'
    TeamResponse:
      type: object
      properties:
        team:
          $ref: '#/components/schemas/TeamSummary'
    TeamSaveRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Team name
    UserLinkRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: string
          description: User identifier
...
paths:
...
  /teams:
    post:
      tags:
        - Teams
      summary: Create or update a team definition
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamSaveRequest'
      responses:
        '201':
          description: Team persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamSummary'
        '400':
          description: Missing fields or invalid payload
        '401':
          description: Missing or invalid token
        '409':
          description: Team with this name alredy exists
        '500':
          description: Internal error
    get:
      tags:
        - Teams
      summary: List teams for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Team list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamListResponse'
        '401':
          description: Missing or invalid token
        '404':
          description: Not found
        '500':
          description: Internal error
  /teams/all:
    get:
      tags:
        - Teams
      summary: List all teams (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Team list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamListResponse'
        '401':
          description: Missing or invalid token
        '403':
          description: Missing admin permission
        '404':
          description: Not found
        '500':
          description: Internal error
  /teams/{id}:
    get:
      tags:
        - Teams
      summary: Retrieve a team by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Team payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller does not belong to the requested team
        '404':
          description: Team not found
        '500':
          description: Internal error
    put:
      tags:
        - Teams
      summary: Update a team definition
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamSaveRequest'
      responses:
        '200':
          description: Team updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Invalid request or payload
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not a team owner or admin
        '404':
          description: Team not found
        '409':
          description: Team with this name already exists
        '500':
          description: Internal error
    delete:
      tags:
        - Teams
      summary: Delete a team
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Team removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not a team owner or admin
        '404':
          description: Team not found
        '500':
          description: Internal error
  /teams/{id}/users:
    post:
      tags:
        - Teams
      summary: Add a user to a team
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLinkRequest'
      responses:
        '200':
          description: Updated team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an owner or admin
        '404':
          description: Team or user not found
  /teams/{id}/users/{userId}:
    delete:
      tags:
        - Teams
      summary: Remove a user from a team
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an owner or admin
        '404':
          description: Team not found
  /teams/{id}/owners:
    post:
      tags:
        - Teams
      summary: Add or promote a team owner
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLinkRequest'
      responses:
        '200':
          description: Updated team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an owner or admin
        '404':
          description: Team or user not found
  /teams/{id}/owners/{userId}:
    delete:
      tags:
        - Teams
      summary: Remove an owner from a team
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an owner or admin
        '404':
          description: Team not found
...
  /users/me/image:
    get:
      tags:
        - Users
      summary: Download the caller's profile picture
      security:
        - bearerAuth: []
      responses:
        '200':
          description: PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Picture not set
    put:
      tags:
        - Users
      summary: Upload or replace the caller's profile picture
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserImageUploadRequest'
      responses:
        '200':
          description: Picture updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid payload
        '401':
          description: Missing or invalid token
        '404':
          description: User not found
    delete:
      tags:
        - Users
      summary: Remove the caller's profile picture
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Picture removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Missing or invalid token
        '404':
          description: User not found
...
  /users/{id}/image:
    get:
      tags:
        - Users
      summary: Download a user's profile picture (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary
        '400':
          description: Missing id
        '404':
          description: Picture or user not found
    put:
      tags:
        - Users
      summary: Upload or replace a user's profile picture (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserImageUploadRequest'
      responses:
        '200':
          description: Picture updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Missing id or invalid payload
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an admin
        '404':
          description: User not found
    delete:
      tags:
        - Users
      summary: Remove a user's profile picture (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Picture removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Missing id
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an admin
        '404':
          description: User not found
...